Refactor folder name companyOfficers to companyofficers
