+++ src/main/java/uk/gov/companieshouse/service/dissolution/validator/DissolutionValidator.java
@@ -2,7 +2,7 @@ package uk.gov.companieshouse.service.dissolution.validator;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import uk.gov.companieshouse.model.dto.companyOfficers.CompanyOfficer;
+import uk.gov.companieshouse.model.dto.companyofficers.CompanyOfficer;
 import uk.gov.companieshouse.model.dto.companyProfile.CompanyProfile;
 import uk.gov.companieshouse.model.dto.dissolution.DirectorRequest;
 import uk.gov.companieshouse.service.CompanyOfficerService;
