+++ src/test/java/uk/gov/companieshouse/service/CompanyOfficerServiceTest.java
@@ -6,7 +6,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.companieshouse.client.CompanyOfficersClient;
-import uk.gov.companieshouse.model.dto.companyOfficers.CompanyOfficer;
+import uk.gov.companieshouse.model.dto.companyofficers.CompanyOfficer;
 import uk.gov.companieshouse.model.dto.dissolution.DirectorRequest;
 import uk.gov.companieshouse.model.enums.OfficerRole;
 import uk.gov.companieshouse.service.dissolution.validator.CompanyOfficerValidator;
